move getWallets module to Wallets service
